Add peak detection to POD periodogram and refactor spectral utils
